+++ src/app/(hasNotMap)/youth-policy/page.module.scss
@@ -1,3 +1,3 @@
 .pageWrapper {
-  margin: 30px 0;
+  margin: 60px 0;
 }
